code updated for the era5.
