+++ src/arco_era5/__init__.py
@@ -34,7 +34,7 @@ from .source_data import (
     _read_nc_dataset
     )
 from .update_ar import UpdateSlice as ARUpdateSlice
-from .update_co import GenerateOffset, UpdateSlice as COUpdateSlice, generate_input_paths
+from .update_co import GenerateOffset, UpdateSlice as COUpdateSlice
 from .update_config_files import (
     get_secret,
     update_date_in_config_file,
@@ -43,14 +43,17 @@ from .update_config_files import (
     get_month_range,
     add_licenses_in_config_files,
     get_last_sixth_date,
+    update_target_path_in_config_file
     )
 from .update_model_level_native_vertical_zarr import LoadDataForDayDoFn, UpdateSlice as UpdateModelLevelNativeVerticalDataSlice, hourly_dates
 from .utils import (
+    convert_to_date,
+    data_splitting_dataflow_job,
     date_range,
+    generate_input_paths,
     replace_non_alphanumeric_with_hyphen,
     subprocess_run,
-    convert_to_date,
     parse_arguments_raw_to_zarr_to_bq,
     raw_data_download_dataflow_job,
-    data_splitting_dataflow_job
+    update_raw_data
     )
