+++ src/arco_era5/update_ar.py
@@ -17,6 +17,7 @@ import logging
 import zarr
 
 import apache_beam as beam
+import numpy as np
 import xarray as xr
 import xarray_beam as xb
 
@@ -49,7 +50,10 @@ class UpdateSlice(beam.PTransform):
         for vname in ds.data_vars:
             logger.info(f"Started {vname} for {date.strftime('%Y-%m-%d')}")
             zv = zf[vname]
-            zv[region] = ds[vname].values
+            ans = np.array_equal(zv[region], ds[vname].values, equal_nan=True)
+            if not ans:
+                logger.info(f"Variable {vname} for {date.strftime('%Y-%m-%d')} not ingested properly.")
+                zv[region] = ds[vname].values
             logger.info(f"Done {vname} for {date.strftime('%Y-%m-%d')}")
         del zv
         del ds
