+++ src/arco_era5/utils.py
@@ -22,8 +22,47 @@ import pandas as pd
 import typing as t
 
 from concurrent.futures import ThreadPoolExecutor
+from itertools import product
+
+from .source_data import (
+    GCP_DIRECTORY,
+    SINGLE_LEVEL_VARIABLES,
+    MULTILEVEL_VARIABLES,
+    PRESSURE_LEVELS_GROUPS
+)
 
 logger = logging.getLogger(__name__)
+
+SINGLE_LEVEL_SUBDIR_TEMPLATE = (
+    "ERA5GRIB/HRES/Month/{year}/{year}{month:02d}_hres_{chunk}.grb2"
+)
+MODELLEVEL_SUBDIR_TEMPLATE = (
+    "ERA5GRIB/HRES/Daily/{year}/{year}{month:02d}{day:02d}_hres_{chunk}.grb2"
+)
+
+# File Templates
+PRESSURELEVEL_DIR_TEMPLATE = (
+    "gs://gcp-public-data-arco-era5/raw/date-variable-pressure_level/{year:04d}/{month:02d}/{day:02d}/{chunk}/{pressure}.nc")
+AR_SINGLELEVEL_DIR_TEMPLATE = (
+    "gs://gcp-public-data-arco-era5/raw/date-variable-single_level/{year:04d}/{month:02d}/{day:02d}/{chunk}/surface.nc")
+
+# Data Chunks
+MODEL_LEVEL_CHUNKS = ["dve", "tw", "o3q", "qrqs"]
+SINGLE_LEVEL_CHUNKS = [
+    "cape", "cisst", "sfc", "tcol", "soil_depthBelowLandLayer_istl1",
+    "soil_depthBelowLandLayer_istl2", "soil_depthBelowLandLayer_istl3",
+    "soil_depthBelowLandLayer_istl4", "soil_depthBelowLandLayer_stl1",
+    "soil_depthBelowLandLayer_stl2", "soil_depthBelowLandLayer_stl3",
+    "soil_depthBelowLandLayer_stl4", "soil_depthBelowLandLayer_swvl1",
+    "soil_depthBelowLandLayer_swvl2", "soil_depthBelowLandLayer_swvl3",
+    "soil_depthBelowLandLayer_swvl4", "soil_surface_tsn", "lnsp",
+    "zs", "rad", "pcp_surface_cp", "pcp_surface_crr",
+    "pcp_surface_csf", "pcp_surface_csfr", "pcp_surface_es",
+    "pcp_surface_lsf", "pcp_surface_lsp", "pcp_surface_lspf",
+    "pcp_surface_lsrr", "pcp_surface_lssfr", "pcp_surface_ptype",
+    "pcp_surface_rsn", "pcp_surface_sd", "pcp_surface_sf",
+    "pcp_surface_smlt", "pcp_surface_tp"]
+PRESSURE_LEVEL = PRESSURE_LEVELS_GROUPS["full_37"]
 
 AR_FILES = ['era5_pl_hourly.cfg', 'era5_sl_hourly.cfg']
 CO_MODEL_LEVEL_FILES = ['era5_ml_dve.cfg', 'era5_ml_o3q.cfg', 'era5_ml_qrqs.cfg', 'era5_ml_tw.cfg']
@@ -126,19 +165,22 @@ def raw_data_download_dataflow_job(python_path: str, project: str, region: str,
                                    manifest_location: str, directory: str,
                                    type: str = None):
     """Launches a Dataflow job to process weather data."""
-    current_day = datetime.date.today()
-    job_name = f"raw-data-download-arco-era5-{current_day.month}-{current_day.year}"
 
     AR_FILES = [ f'{directory}/{file}' for file in AR_FILES ]
     CO_MODEL_LEVEL_FILES = [ f'{directory}/{file}' for file in CO_MODEL_LEVEL_FILES ]
     CO_SINGLE_LEVEL_FILES = [ f'{directory}/{file}' for file in CO_SINGLE_LEVEL_FILES ]
+    
+    current_day = datetime.date.today()
 
     if type == 'ERA5T_DAILY':
         files = ' '.join(AR_FILES + CO_MODEL_LEVEL_FILES)
+        job_name = f"raw-data-download-arco-era5T-Daily-{current_day}"
     elif type == 'ERA5T_MONTHLY':
         files = ' '.join(CO_SINGLE_LEVEL_FILES)
+        job_name = f"raw-data-download-arco-era5T-Monthly-{current_day.month}-{current_day.year}"
     else:
         files = ' '.join(AR_FILES + CO_MODEL_LEVEL_FILES + CO_SINGLE_LEVEL_FILES)
+        job_name = f"raw-data-download-arco-era5-{current_day.month}-{current_day.year}"
     command = (
         f"{python_path} /weather/weather_dl/weather-dl {files} "
         f"--runner DataflowRunner --project {project} --region {region} --temp_location "
@@ -174,3 +216,79 @@ def data_splitting_dataflow_job(python_path: str, project: str, region: str,
     with ThreadPoolExecutor(max_workers=4) as tp:
         for command in commands:
             tp.submit(subprocess_run, command)
+
+
+def generate_urls(data_date_range: t.List[datetime.datetime], start_date: str, end_date: str, type: str = None) -> t.List['str']:
+    all_uri = []
+    if type != 'ERA5T_MONTHLY':
+        all_uri.extend(generate_input_paths(start_date, end_date, GCP_DIRECTORY, MODEL_LEVEL_CHUNKS))
+        for date in data_date_range:
+            for chunk in MULTILEVEL_VARIABLES + SINGLE_LEVEL_VARIABLES:
+                if chunk in MULTILEVEL_VARIABLES:
+                    for pressure in PRESSURE_LEVEL:
+                        all_uri.append(
+                            PRESSURELEVEL_DIR_TEMPLATE.format(year=date.year,
+                                                            month=date.month,
+                                                            day=date.day, chunk=chunk,
+                                                            pressure=pressure))
+                else:
+                    if chunk == 'geopotential_at_surface':
+                        chunk = 'geopotential'
+                    all_uri.append(
+                        AR_SINGLELEVEL_DIR_TEMPLATE.format(
+                            year=date.year, month=date.month, day=date.day, chunk=chunk))
+
+    if type != 'ERA5T_DAILY':   
+        all_uri.extend(generate_input_paths(start_date, end_date, GCP_DIRECTORY, SINGLE_LEVEL_CHUNKS, True))
+    
+    return all_uri
+
+
+def update_raw_data(data_date_range: t.List[datetime.datetime]):
+    """
+    Copies contents from src to dst in Google Cloud Storage, overriding existing files.
+
+    Parameters:
+    - src (str): Source directory URI (e.g., 'gs://bucket/source-dir/')
+    - dst (str): Destination directory URI (e.g., 'gs://bucket/destination-dir/')
+    """
+    start_date = data_date_range[0].strftime("%Y/%m/%d")
+    end_date = data_date_range[-1].strftime("%Y/%m/%d")
+
+    all_uri = generate_urls(data_date_range, start_date, end_date, type)
+
+    with ThreadPoolExecutor() as tp:
+        for url in all_uri:
+            src = url.replace("gcp-public-data-arco-era5/raw", "gcp-public-data-arco-era5/raw-era5")
+            command = f'gsutil -m mv {src} {url}'
+            tp.submit(subprocess_run, command)
+
+
+def generate_input_paths(start: str, end: str, root_path: str, chunks: t.List[str],
+                         is_single_level: bool = False) -> t.List[str]:
+    """A method for generating the url using the combination of chunks and time range.
+
+    Args:
+        start (str): starting date for data files.
+        end (str): last date for data files.
+        root_path (str): raw file url prefix.
+        chunks: (t.List(str)): List of chunks to process.
+        is_single_level: (bool): Is the call for single level or model level
+
+    Returns:
+        t.List: List of file urls to process.
+    """
+    input_paths = []
+    for time, chunk in product(date_range(start, end, freq="MS" if is_single_level else "D"), chunks):
+        if is_single_level:
+            url = f"{root_path}/{SINGLE_LEVEL_SUBDIR_TEMPLATE.format(year=time.year, month=time.month, day=time.day, chunk=chunk)}"
+        else:
+            url = f"{root_path}/{MODELLEVEL_SUBDIR_TEMPLATE.format(year=time.year, month=time.month, day=time.day, chunk=chunk)}"
+
+        if '_' in chunk:
+            chunk_, level, var = chunk.split('_')
+            url = url.replace(chunk, chunk_)
+            url = f"{url}_{level}_{var}.grib"
+        input_paths.append(url)
+
+    return input_paths
